style: Assert over frame counts without crossing signedness (#682)
reached_k_ is an int that starts at -1, but each of these asserts sits
past a guard that has already raised it, so the frame count it is
compared against is the same kind of quantity.

+++ engines/dual_approx_reach.cpp
@@ -144,7 +144,7 @@ bool DualApproxReach::step(int i)
 
   ++reached_k_;
   assert(i == reached_k_);
-  assert(forward_seq_.size() == reached_k_ + 1);
+  assert(forward_seq_.size() == static_cast<size_t>(reached_k_) + 1);
   assert(forward_seq_.size() == backward_seq_.size());
 
   return check_fixed_point();
